server: fix heartbeat stream bug. (#709)
