+++ server/coordinator_test.go
@@ -49,7 +49,9 @@ type testCoordinatorSuite struct{}
 func (s *testCoordinatorSuite) TestBasic(c *C) {
 	cluster := newClusterInfo(newMockIDAllocator())
 	_, opt := newTestScheduleConfig()
-	co := newCoordinator(cluster, opt)
+	hbStreams := newHeartbeatStreams(cluster.getClusterID())
+	defer hbStreams.Close()
+	co := newCoordinator(cluster, opt, hbStreams)
 	l := co.limiter
 
 	op1 := newTestOperator(1, LeaderKind)
@@ -96,9 +98,11 @@ func newMockHeartbeatStream() *mockHeartbeatStream {
 func (s *testCoordinatorSuite) TestDispatch(c *C) {
 	cluster := newClusterInfo(newMockIDAllocator())
 	tc := newTestClusterInfo(cluster)
+	hbStreams := newHeartbeatStreams(cluster.getClusterID())
+	defer hbStreams.Close()
 
 	_, opt := newTestScheduleConfig()
-	co := newCoordinator(cluster, opt)
+	co := newCoordinator(cluster, opt, hbStreams)
 	co.run()
 	defer co.stop()
 
@@ -158,13 +162,15 @@ func dispatchAndRecvHeartbeat(co *coordinator, region *RegionInfo, stream *mockH
 func (s *testCoordinatorSuite) TestReplica(c *C) {
 	cluster := newClusterInfo(newMockIDAllocator())
 	tc := newTestClusterInfo(cluster)
+	hbStreams := newHeartbeatStreams(cluster.getClusterID())
+	defer hbStreams.Close()
 
 	// Turn off balance.
 	cfg, opt := newTestScheduleConfig()
 	cfg.LeaderScheduleLimit = 0
 	cfg.RegionScheduleLimit = 0
 
-	co := newCoordinator(cluster, opt)
+	co := newCoordinator(cluster, opt, hbStreams)
 	co.run()
 	defer co.stop()
 
@@ -214,9 +220,11 @@ func (s *testCoordinatorSuite) TestReplica(c *C) {
 func (s *testCoordinatorSuite) TestPeerState(c *C) {
 	cluster := newClusterInfo(newMockIDAllocator())
 	tc := newTestClusterInfo(cluster)
+	hbStreams := newHeartbeatStreams(cluster.getClusterID())
+	defer hbStreams.Close()
 
 	_, opt := newTestScheduleConfig()
-	co := newCoordinator(cluster, opt)
+	co := newCoordinator(cluster, opt, hbStreams)
 	co.run()
 	defer co.stop()
 
@@ -261,9 +269,11 @@ func (s *testCoordinatorSuite) TestPeerState(c *C) {
 func (s *testCoordinatorSuite) TestShouldRun(c *C) {
 	cluster := newClusterInfo(newMockIDAllocator())
 	tc := newTestClusterInfo(cluster)
+	hbStreams := newHeartbeatStreams(cluster.getClusterID())
+	defer hbStreams.Close()
 
 	_, opt := newTestScheduleConfig()
-	co := newCoordinator(cluster, opt)
+	co := newCoordinator(cluster, opt, hbStreams)
 
 	tc.LoadRegion(1, 1, 2, 3)
 	tc.LoadRegion(2, 1, 2, 3)
@@ -294,11 +304,13 @@ func (s *testCoordinatorSuite) TestShouldRun(c *C) {
 func (s *testCoordinatorSuite) TestAddScheduler(c *C) {
 	cluster := newClusterInfo(newMockIDAllocator())
 	tc := newTestClusterInfo(cluster)
+	hbStreams := newHeartbeatStreams(cluster.getClusterID())
+	defer hbStreams.Close()
 
 	cfg, opt := newTestScheduleConfig()
 	cfg.ReplicaScheduleLimit = 0
 
-	co := newCoordinator(cluster, opt)
+	co := newCoordinator(cluster, opt, hbStreams)
 	co.run()
 	defer co.stop()
 
@@ -348,6 +360,42 @@ func (s *testCoordinatorSuite) TestAddScheduler(c *C) {
 	c.Assert(resp, IsNil)
 }
 
+func (s *testCoordinatorSuite) TestRestart(c *C) {
+	cluster := newClusterInfo(newMockIDAllocator())
+	tc := newTestClusterInfo(cluster)
+	hbStreams := newHeartbeatStreams(cluster.getClusterID())
+	defer hbStreams.Close()
+
+	// Turn off balance, we test add replica only.
+	cfg, opt := newTestScheduleConfig()
+	cfg.LeaderScheduleLimit = 0
+	cfg.RegionScheduleLimit = 0
+
+	// Add 3 stores (1, 2, 3) and a region with 1 replica on store 1.
+	tc.addRegionStore(1, 1)
+	tc.addRegionStore(2, 2)
+	tc.addRegionStore(3, 3)
+	tc.addLeaderRegion(1, 1)
+	cluster.activeRegions = 1
+	region := cluster.getRegion(1)
+
+	// Add 1 replica on store 2.
+	co := newCoordinator(cluster, opt, hbStreams)
+	co.run()
+	stream := newMockHeartbeatStream()
+	resp := dispatchAndRecvHeartbeat(co, region, stream)
+	checkAddPeerResp(c, resp, 2)
+	region.Peers = append(region.Peers, resp.GetChangePeer().GetPeer())
+	co.stop()
+
+	// Recreate coodinator then add another replica on store 3.
+	co = newCoordinator(cluster, opt, hbStreams)
+	co.run()
+	resp = dispatchAndRecvHeartbeat(co, region, stream)
+	checkAddPeerResp(c, resp, 3)
+	co.stop()
+}
+
 func waitOperator(c *C, co *coordinator, regionID uint64) {
 	for i := 0; i < 20; i++ {
 		if co.getOperator(regionID) != nil {
@@ -391,7 +439,9 @@ type testScheduleControllerSuite struct{}
 func (s *testScheduleControllerSuite) TestController(c *C) {
 	cluster := newClusterInfo(newMockIDAllocator())
 	cfg, opt := newTestScheduleConfig()
-	co := newCoordinator(cluster, opt)
+	hbStreams := newHeartbeatStreams(cluster.getClusterID())
+	defer hbStreams.Close()
+	co := newCoordinator(cluster, opt, hbStreams)
 	lb := newBalanceLeaderScheduler(opt)
 	sc := newScheduleController(co, lb, minScheduleInterval)
 
@@ -442,7 +492,9 @@ func (s *testScheduleControllerSuite) TestController(c *C) {
 func (s *testScheduleControllerSuite) TestInterval(c *C) {
 	cluster := newClusterInfo(newMockIDAllocator())
 	_, opt := newTestScheduleConfig()
-	co := newCoordinator(cluster, opt)
+	hbStreams := newHeartbeatStreams(cluster.getClusterID())
+	defer hbStreams.Close()
+	co := newCoordinator(cluster, opt, hbStreams)
 	lb := newBalanceLeaderScheduler(opt)
 	sc := newScheduleController(co, lb, minScheduleInterval)
